requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
ollama==0.4.7
cloudinary==1.41.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
